requirements.txt
defusedxml==0.7.1
detectron2==0.6
dill==0.3.8
matplotlib==3.9.3
mmrotate==0.3.4
numpy==2.2.0
opencv_python==4.10.0.84
packaging==24.2
paddle==1.0.2
Pillow==11.0.0
prettytable==3.12.0
pybind11==2.13.4
pycocotools==2.0.8
pydicom==3.0.1
PyYAML==6.0.1
PyYAML==6.0.2
rich==13.9.4
scikit_learn==1.6.0
scipy==1.14.1
setuptools==60.2.0
Shapely==2.0.6
sympy==1.12
timm==1.0.12
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.2
utiles==0.6.1
